Add files via upload
I realized one mistake in my answer to question 1 while doing question 3 - the interest rate at which today you pay back yesterday's debt only depend on today's k,b, and yesterday's a(productivity). I made the mistake of making it depend on k' and b'. This is all related to how to decide on RbMinus.